Reenable building svgSalamander
The removal of the netbeans project files broke the build of svgSalamander.
To fix this, the build was migrated to maven. The maven project can be directly
opened by netbeans and build from command line. The project can now be build:

cd svg-core
mvn package

will create a file svgSalamander-<VERSION>.jar in the target directory.

The build can also be used to directly install the artifact into the local
repository:

cd svg-core
mvn install

This is necessary to use the examples.