Properly handle all possible Graphics transforms
We previously assumed that all possible transforms have a
scale component which is positive. Even for a transform as
simple as a 90 degree rotation this is not true and would
break the painting method.
We now correctly paint the element to a pre-transformed
buffer which can be blitted correctly.
Notably this also fixes picking with non scale or translate
transforms.