Implement Masks (#78)
* Ignore intelliJ project files

* Allow for rendering mixins

Move rendering into the RenderableElement#doRender method, which gets called by RenderableElement#render. This allows for mixins while rendering.

* Allow for pick mixins

As for rendering we want to allow intercepting calls to pick in RenderableElement.

* Implement Mask element

If a mask is present it takes over painting of the element. This is done by first painting the element to a buffer applying the mask and the blitting the image. For now this assumes that the mask uses a 'color-interpolation' value of 'linearRGB'.

* Ensure the Graphics scale is set to (1.0, 1.0) when painting the masked image

* Cache the Mask element in `build`.

This avoids fetching it in `render` and `pick` each time.

* Properly handle all possible Graphics transforms

We previously assumed that all possible transforms have a
scale component which is positive. Even for a transform as
simple as a 90 degree rotation this is not true and would
break the painting method.
We now correctly paint the element to a pre-transformed
buffer which can be blitted correctly.
Notably this also fixes picking with non scale or translate
transforms.