Provide infrastructure for filters
Because filters like mask need to facilitate painting to an off-screen buffer,
we extract the logic for this into a separate class outside of Mask.
Then the size of the buffer is first adjusted to suit the installed filters.
After applying the filters the mask is painted as usual.
Notably this implementation currently doesn't support any filter specific
attributes such as 'in', 'feOffset', 'feMergeMode' etc.