Implement rudimentary gaussian blur (#82)
* Make path parsing less error prone

Replace the regex based path parser with a simpler version.

* Provide infrastructure for filters

Because filters like mask need to facilitate painting to an off-screen buffer,
we extract the logic for this into a separate class outside of Mask.
Then the size of the buffer is first adjusted to suit the installed filters.
After applying the filters the mask is painted as usual.
Notably this implementation currently doesn't support any filter specific
attributes such as 'in', 'feOffset', 'feMergeMode' etc.

* Implement gaussian blur filter effect

This doesn't yet do the recommended optimization for standard deviations larger
than 2, replacing the gaussian kernel by successive box blurs.

* Make target component available during painting

If the diagram is painted to a JComponent make it available s.t. it can be used
as the ImageObserver, when painting images. Otherwise they may not be fully loaded
when painting.

* Cache the off-screen buffer

This avoids unnecessary calculations when presenting the svg in a GUI.
We only do this if we are painting to a component as it wastes memory otherwise.